Commit progress on lilim parser

+++ src/utils/routing.rs
@@ -2,7 +2,7 @@ use crate::models::common::Location;
 use crate::models::problem::MatrixTransportCost;
 
 pub struct MatrixFactory {
-    locations: Vec<(usize, usize)>,
+    locations: Vec<(i32, i32)>,
 }
 
 impl MatrixFactory {
@@ -10,7 +10,7 @@ impl MatrixFactory {
         Self { locations: vec![] }
     }
 
-    pub fn collect(&mut self, location: (usize, usize)) -> Location {
+    pub fn collect(&mut self, location: (i32, i32)) -> Location {
         match self.locations.iter().position(|l| l.0 == location.0 && l.1 == location.1) {
             Some(position) => position,
             _ => {
